basic "move cursor down"
still buggy, but the core is there. It really does not work nice if the
font is proportional because it uses text offsets instead of pixel
distances.

We need to decide what kind of editor we are building,  to take a pick
here.

Also got tired of CRLF issues so now I filter them out at loading time.