add a scale factor
ctrl + mousewheel zooms in our out and a scale_ variable holds the
current scale.  The scale is applied with center at (width/2, 0).

the title gets updated to show the current scale. which is applied
pretty much to everything except the title piece and the widgets.

The cursor hit testing with mouse is not updated yet and there is a hack
for |v_max| so we render more.