add EOF textblock
We add a final textblock with a special state == eof and make the
needs_layout() properly reflect the fact that we have a dwrite layout
object or not. Previously being smuggled on the metrics.lineCount
member.

This fixes a problem that the cursor could not be moved by mouse to the
last block because the last block would never be in the loop.

Same problem if instead of loading a document the user just typed a few
words, the cursor could not be positioned with the mouse either.