fix bug when zoomed in
when zoomed in the last block visible was not mouse hit testable for
cursor.

Just another variation of last_block_in_view_ issues. Therefore I
decided to do away with that variable and just loop until the end of the
textblock array if necessary.

The EOF block is still necessary since the logic still backtracks to the
previous block  as in "text_[ix].metrics.top > y" needs to advance to
the EOF block.