caret line improvements
we now draw the cursor line (like visual studio) and we scale the caret
up or down so it is visible at all supported scales.

Technically it would be better to draw the cursor line first and then
the text. Since we draw the line after the text we are constrained in
color, or else the text would look washed.