redo all text management
Remove TextBlock and Cursor classes from the code. TExTO is now fully
broken, but still paints the basic UI.

long story short, got tired of all the crappy code related to managing
the TextBlocks, so I decided to try door #2, which is a single block of
text, approximately 8k of text that is the 'visible buffer' and the rest
is just kept a plain string.

If this works, lots of things are going to be much simpler. The
drawback, which led me not to pursue this was the copying back and forth
from the working buffer to the string.