redo text management part 6
v_scroll is gone, we now use active_start_ to indicate what line we are
in, and so when we scroll we jump to per line instead of pixels.

To scroll forward, we use DWRITE_LINE_METRICS and to scroll back we use
std::vector<size_t> starts_.

If there is no active text (now a unique_ptr) we simply use the pointers
into the full text.