remove silly methods
set_full_text goes away and now the program needs to delete the existing
TextView if it wants to load a file. The full text from the file is just
part of the TextView ctor.

This makes thing easier to reason about.

Also set_text_format becomes part of the ctor, you need that always
anyhow.